LayoutTests/fast/history/page-cache-active-fetch-request-blobReadAsText-expected.txt

Tests that a page with an active fetch request is able to enter the back/forward cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the back/forward cache
PASS Promise resolved after restoring from the back/forward cache.
PASS text is "\"This is the body\""
PASS successfullyParsed is true

TEST COMPLETE
